refactor(test:revision): replace mock_user with mock_user_service
