+++ tests/app/api_v0/test_revisions.py
@@ -1,17 +1,9 @@
 # TODO: split E2E test to unit test
-from typing import Dict, Iterator
 
 import pytest
-from sqlalchemy.orm import Session
 from starlette.testclient import TestClient
 
-import pol.server
-from pol.models import Avatar, PublicUser
-from pol.db.tables import ChiiRevHistory
-from tests.conftest import MockUser
-from pol.services.user_service import UserService
-from pol.services.rev_service.person_rev import person_rev_type_filters
-from pol.services.rev_service.character_rev import character_rev_type_filters
+from tests.conftest import MockUserService
 
 person_revisions_api_prefix = "/v0/revisions/persons"
 
@@ -19,13 +11,8 @@ person_revisions_api_prefix = "/v0/revisions/persons"
 @pytest.mark.env("e2e", "database")
 def test_person_revisions_basic(
     client: TestClient,
-    db_session: Session,
-    mock_user: MockUser,
+    mock_user_service: MockUserService,
 ):
-    for r in db_session.query(ChiiRevHistory.rev_creator).where(
-        ChiiRevHistory.rev_mid == 9, person_rev_type_filters
-    ):
-        mock_user(r.rev_creator)
     response = client.get(person_revisions_api_prefix, params={"person_id": 9})
     assert response.status_code == 200
     assert response.headers["content-type"] == "application/json"
@@ -42,13 +29,8 @@ def test_person_revisions_basic(
 @pytest.mark.env("e2e", "database")
 def test_person_revisions_offset(
     client: TestClient,
-    db_session: Session,
-    mock_user: MockUser,
+    mock_user_service: MockUserService,
 ):
-    for r in db_session.query(ChiiRevHistory.rev_creator).where(
-        ChiiRevHistory.rev_mid == 9, person_rev_type_filters
-    ):
-        mock_user(r.rev_creator)
     offset = 1
     common_params = {"person_id": 9}
     response1 = client.get(
@@ -70,13 +52,8 @@ def test_person_revisions_offset(
 @pytest.mark.env("e2e", "database")
 def test_person_revisions_offset_limit(
     client: TestClient,
-    db_session: Session,
-    mock_user: MockUser,
+    mock_user_service: MockUserService,
 ):
-    for r in db_session.query(ChiiRevHistory.rev_creator).where(
-        ChiiRevHistory.rev_mid == 9, person_rev_type_filters
-    ):
-        mock_user(r.rev_creator)
     offset = 30000
     response = client.get(
         person_revisions_api_prefix, params={"offset": offset, "person_id": 9}
@@ -90,13 +67,8 @@ character_revisions_api_prefix = "/v0/revisions/characters"
 @pytest.mark.env("e2e", "database")
 def test_character_revisions_basic(
     client: TestClient,
-    db_session: Session,
-    mock_user: MockUser,
+    mock_user_service: MockUserService,
 ):
-    for r in db_session.query(ChiiRevHistory.rev_creator).where(
-        ChiiRevHistory.rev_mid == 1, character_rev_type_filters
-    ):
-        mock_user(r.rev_creator)
     response = client.get(character_revisions_api_prefix, params={"character_id": 1})
     assert response.status_code == 200, response.json()
     assert response.headers["content-type"] == "application/json"
@@ -111,13 +83,8 @@ def test_character_revisions_basic(
 @pytest.mark.env("e2e", "database")
 def test_character_revisions_offset(
     client: TestClient,
-    db_session: Session,
-    mock_user: MockUser,
+    mock_user_service: MockUserService,
 ):
-    for r in db_session.query(ChiiRevHistory.rev_creator).where(
-        ChiiRevHistory.rev_mid == 1, character_rev_type_filters
-    ):
-        mock_user(r.rev_creator)
     offset = 1
     common_params = {"character_id": 1}
     response1 = client.get(
@@ -139,13 +106,8 @@ def test_character_revisions_offset(
 @pytest.mark.env("e2e", "database")
 def test_character_revisions_page_limit(
     client: TestClient,
-    db_session: Session,
-    mock_user: MockUser,
+    mock_user_service: MockUserService,
 ):
-    for r in db_session.query(ChiiRevHistory.rev_creator).where(
-        ChiiRevHistory.rev_mid == 1, character_rev_type_filters
-    ):
-        mock_user(r.rev_creator)
     offset = 30000
     response = client.get(
         character_revisions_api_prefix, params={"character_id": 1, "offset": offset}
@@ -156,35 +118,10 @@ def test_character_revisions_page_limit(
 subject_revisions_api_prefix = "/v0/revisions/subjects"
 
 
-class MockUserService:
-    async def get_users_by_id(self, ids: Iterator[int]) -> Dict[int, PublicUser]:
-        ids = list(ids)
-        for uid in ids:
-            assert uid > 0
-
-        return {
-            uid: PublicUser(
-                id=uid,
-                username=f"username {uid}",
-                nickname=f"nickname {uid}",
-                avatar=Avatar.from_db_record(""),
-            )
-            for uid in ids
-        }
-
-    async def get_by_uid(self, uid: int) -> PublicUser:
-        assert uid > 0
-        return PublicUser(
-            id=uid,
-            username=f"username {uid}",
-            nickname=f"nickname {uid}",
-            avatar=Avatar.from_db_record(""),
-        )
-
-
 @pytest.mark.env("e2e", "database")
-def test_subject_revisions_basic(client: TestClient):
-    pol.server.app.dependency_overrides[UserService.new] = MockUserService
+def test_subject_revisions_basic(
+    client: TestClient, mock_user_service: MockUserService
+):
     response = client.get(subject_revisions_api_prefix, params={"subject_id": 26})
     assert response.status_code == 200
     assert response.headers["content-type"] == "application/json"
